% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.plottables.R
\name{add.plottables}
\alias{add.plottables}
\title{Attach plottable variables to a dose-titration study}
\usage{
add.plottables(de, tox.pending = FALSE)
}
\arguments{
\item{de}{A data frame describing a dose-titration study.}

\item{tox.pending}{Set TRUE to show last-period DLT assessments pending}
}
\value{
A data frame derived from \code{de}, with additional columns as
follows: \item{cohort}{Factor identifying each participant's cohort}
\item{x}{Horizontal position at which to plot participant on OXplot}
\item{pch}{Integer ASCII code for participant's plotting character (per
period)}
}
\description{
Prepare a dose-titration study data frame for plotting.
}
\seealso{
\code{\link{OXplot}}
}
\author{
David C. Norris
}
\keyword{hplot}
