% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{methods}
\name{as_d3_data,DE-method}
\alias{as_d3_data,DE-method}
\title{Convert a DE object to JSON}
\usage{
\S4method{as_d3_data}{DE}(x, ...)
}
\arguments{
\item{x}{An object of S4 class 'DE'}

\item{\dots}{Unused; included to match signature of generic method}
}
\description{
Convert a DE object to JSON
}
