% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_d3_data.de.R
\name{as_d3_data}
\alias{as_d3_data.de}
\title{Transform the 'de' object to JSON data for D3 visualization.}
\usage{
\method{as_d3_data}{de}(x, ...)
}
\arguments{
\item{x}{The 'de' object to be transformed}

\item{\dots}{Unused; included to match signature of generic method}
}
\value{
The returned value is a JSON string corresponding to a list of data
frames: \item{mtd}{Data frame with columns c('id','mtd')} \item{trial}{Data
frame with 'id', 'period', 'dose', 'dlt'.)} \item{ds}{Data frame with
columns 'period','surv','lower','upper','dose'.}
}
\description{
The returned object is a JSON representation of a JavaScript object having
components 'mtd', 'trial' and 'ds' -- the last of which stands for
'dose-survival'.
}
\details{
The (unhelpful) default behavior of function \code{r2d3} is to effect this
transformation so as to reduce the size of transmitted data, then restore
the original form upon receipt using JavaScript method
\code{HTMLWidgets.dataframeToD3()}. See
\url{https://rstudio.github.io/r2d3/articles/data_conversion.html}.
}
\examples{

library(DTAT)
de <- de.sim(testing=TRUE)
data <- as_d3_data(de)
## Read 'data' back into a list and check a few things
check <- jsonlite::fromJSON(data)
stopifnot(max(abs(attr(de,'mtd') - check$mtd)) < 1e-4)
stopifnot(max(abs(attr(de,'doses') - check$doses)) < 1e-4)
stopifnot(max(abs(attr(de,'mtd_quantiles') - check$mtd_quantiles)) < 1e-4)
stopifnot(all(de[[length(de)]] - check$trial == 0))

}
\author{
David C. Norris
}
\keyword{dynamic}
\keyword{manip}
