% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble}
\alias{ensemble}
\title{Generate and summarize an ensemble of simulated dose titration studies.}
\usage{
ensemble(K = 10, ...)
}
\arguments{
\item{K}{Number of simulated studies to run}

\item{\dots}{Parameters to be passed through to \code{de.sim}}
}
\value{
A data frame summarizing key outcomes of the simulated trials:
\item{enrolling.dose}{Enrolling dose at conclusion of study}
\item{max.dose}{Maximum allowed dose at conclusion of study}
\item{effic.33pc}{Efficiency of titration over retained doses, relative to
perfect 'MTDi' dosing} \item{effic.1sfa}{Efficiency of 1-size-fits-all
dosing, relative to perfect 'MTDi' dosing}
}
\description{
Generate and summarize an ensemble of simulated dose titration studies.
}
\examples{

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (K = 10, ...) 
{
    pb <- txtProgressBar(max = K, style = 3)
    df <- de.sim(...)
    MTDi.ensemble <<- list(MTDi.last)
    de.ensemble <<- list(de.last)
    setTxtProgressBar(pb, 1)
    for (k in 2:K) {
        df <- rbind(df, de.sim(...))
        MTDi.ensemble[[k]] <<- list(MTDi.last)
        de.ensemble[[k]] <<- de.last
        setTxtProgressBar(pb, k)
    }
    df
  }

}
\seealso{
\code{\link{de.sim}}
}
\author{
David C. Norris
}
\keyword{~kwd1}
\keyword{~kwd2}
