% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OXDSanimate.R
\name{OXDSanimate}
\alias{OXDSanimate}
\title{Transparent animation of a 3+3/PC dose-titration study}
\usage{
OXDSanimate(de)
}
\arguments{
\item{de}{A list of period-wise snapshots from a dose-titration study}
}
\value{
No value is returned. This function is intended to be passed as the
\code{expr} argument to package \code{animation} functions such as
\code{saveGIF} and \code{saveVideo}. See the example below.
}
\description{
Animate the periods of a 3+3/PC dose-titration study, showing the
dose-survival curve as it accumulates with the conclusion of each DLT
assessment period.
}
\examples{

\dontrun{
# Running this is not quite 'catastrophic', but the saveVideo()
# function may take a good fraction of a minute to run.
library(animation)
ani.options(ani.width=1012, ani.height=506)
de <- de.sim(testing=TRUE)
saveVideo(OXDSanimate(de)) # produces file "animation.mp4"
}

}
\seealso{
\code{\link{OXDSplot}} \code{\link[animation]{animation}}
}
\author{
David C. Norris
}
\keyword{dynamic}
