% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OXDSplot.R
\name{OXDSplot}
\alias{OXDSplot}
\title{Transparent visualization of a 3+3/PC dose-titration study}
\usage{
OXDSplot(de, tox.pending = FALSE, periods = max(de$period),
  top.dose = max(de$dose), ox.width = 0.7)
}
\arguments{
\item{de}{A data frame describing a dose-titration study}

\item{tox.pending}{Set TRUE to show last-period DLT assessment pending}

\item{periods}{Number of periods of \code{de} to show}

\item{top.dose}{Highest dose level to show on the plot}

\item{ox.width}{Fraction of plot width occupied by the 'OX plot'}
}
\value{
A \code{lattice} plot
}
\description{
Visualize a 3+3/PC dose-titration study
}
\details{
Figure 1 of the reference below was generated by this function. Please see
the extended caption of that figure for details on interpretation of this
complex plot.
}
\references{
Norris DC. Precautionary Coherence Unravels Dose Escalation
Designs. bioRxiv. December 2017:240846. doi:10.1101/240846.
\url{https://www.biorxiv.org/content/early/2017/12/29/240846}
}
\seealso{
\code{\link{OXplot}}, \code{\link{xyplot.survfit}}
}
\author{
David C. Norris
}
\keyword{hplot}
\keyword{survival}
