% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OXplot.R
\name{OXplot}
\alias{OXplot}
\title{Visualize the 'percolation' through the dose levels of a 3+3/PC titration
study}
\usage{
OXplot(de, tox.pending = FALSE, periods = max(de$period),
  top.dose = max(de$dose), ox.col = brewer.pal(4, "Dark2"))
}
\arguments{
\item{de}{A data frame describing a dose-titration study}

\item{tox.pending}{Set TRUE to show last-period DLT assessment pending}

\item{periods}{Number of periods of \code{de} to show}

\item{top.dose}{Highest dose level to show on the plot}

\item{ox.col}{Colors to cycle through to help discriminate enrolled cohorts}
}
\value{
A \code{lattice} plot
}
\description{
Visualize participants' 'percolation' through the dose levels of a 3+3/PC
titration study
}
\references{
Norris DC. Precautionary Coherence Unravels Dose Escalation
Designs. bioRxiv. December 2017:240846. doi:10.1101/240846.
\url{https://www.biorxiv.org/content/early/2017/12/29/240846}
}
\seealso{
\code{\link{OXDSplot}} uses \code{OXplot} in its full and proper
context; please consult its documentation for full details
}
\author{
David C. Norris
}
\keyword{hplot}
