% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramset.R
\name{paramset}
\alias{paramset}
\title{Set up a parameter vector from sim$pop[id,]}
\usage{
paramset(id, states = NULL, Tinfusion = 1, dose1 = 50)
}
\arguments{
\item{id}{Subject identifier, an index into \code{pop} global variable.}

\item{states}{This looks like a 'legacy' parameter, no longer used. TODO:
Factor it out?}

\item{Tinfusion}{Duration of infusion, in hours.}

\item{dose1}{Infusion dose.}
}
\value{
A named parameter vector, suitable a the \code{params} argument for
\code{pomp::trajectory}
}
\description{
Obtain a parameter vector for given id from global \code{sim$pop} data
frame.
}
\details{
TODO: This is really a straightforward utility function, apparently used only
once, in function \code{titrate}. Perhaps I should have made it an internal
function!
}
\seealso{
\code{\link[pomp]{trajectory}}
}
\author{
David C. Norris
}
