% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{methods}
\name{plot,DE,missing-method}
\alias{plot,DE,missing-method}
\title{Plot a DE object as an interactive htmlwidget}
\usage{
\S4method{plot}{DE,missing}(x, y, ...)
}
\arguments{
\item{x}{An object of class \code{DE}}

\item{y}{Unused; included for S4 generic consistency}

\item{...}{Passed to \code{\link[r2d3]{r2d3}}, enabling caller to (e.g.) the
override the default \code{viewer = "internal"}.}
}
\description{
Plot a DE object as an interactive htmlwidget
}
