% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_d3_data}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages.
Follow the linksbelow to see their documentation.
\itemize{
\item r2d3 \code{\link[r2d3]{as_d3_data}}
}
}
\keyword{internal}
