% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDTATapp.R
\name{runDTATapp}
\alias{runDTATapp}
\title{Run Shiny apps included in package:DTAT}
\usage{
runDTATapp(app)
}
\arguments{
\item{app}{Character vector of length 1. Name off app to run.}
}
\value{
Invoked for side effect. Runs the named Shiny app.
}
\description{
Run Shiny apps included in package:DTAT
}
\examples{
\donttest{
runApp("Sim33PC")
}
}
