\name{scaled}
\alias{scaled}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Power-law scaling for doses
}
\description{
Implement an inverse power-law scaling for drug dose. 
}
\usage{
scaled(dose, a = 4)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dose}{A numeric vector of doses}
  \item{a}{A numeric exponent for power-law rescaling}
}
\value{A rescaled vector of doses}
\author{David C. Norris}
