% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{test.DE}
\alias{test.DE}
\title{Temporary test of 'DE' class}
\usage{
test.DE(seed = 2017, CV = 0.7, mean_mtd = 1, start.dose = 0.25,
  dose.jump = 0.4)
}
\arguments{
\item{seed}{RNG seed}

\item{CV}{Coefficient of variation of Gamma-distributed MTDi}

\item{mean_mtd}{Mean of Gamma-distributed MTDi}

\item{start.dose}{Lowest (starting) dose to trial}

\item{dose.jump}{Proportional step between geometrically-spaced doses}
}
\description{
TODO: Implement a 'sim.DE'-like function and a more comprehensive
suite of tests.
}
