% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{methods}
\name{titration}
\alias{titration}
\alias{titration,DE,numeric-method}
\title{Simulate a \sQuote{3+3/PC} dose-titration trial}
\usage{
titration(x, periods, ...)

\S4method{titration}{DE,numeric}(x, periods, ...)
}
\arguments{
\item{x}{An object of S4 class \code{\link[=DE-class]{DE}}}

\item{periods}{The number of DLT assessment periods to titrate over.
Should be a positive integer.}

\item{...}{May be used to pass \code{verbatim = 'TRUE'} to internal
\code{step_time} method.}
}
\description{
Simulate a \sQuote{3+3/PC} dose-titration trial
}
\references{
Norris DC. Precautionary Coherence Unravels Dose Escalation Designs.
\emph{bioRxiv}. December 2017:240846. doi:10.1101/240846.
\url{https://www.biorxiv.org/content/early/2017/12/29/240846}
}
