% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyplot.survfit.R
\name{xyplot.survfit}
\alias{xyplot.survfit}
\title{Plot a dose-survival curve with its confidence bounds}
\usage{
\method{xyplot}{survfit}(x, data = NULL, ylim = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{survfit}}

\item{data}{Unused; included for S3 generic/method consistency.}

\item{ylim}{Passed through to xyplot}

\item{\dots}{Passed through to xyplot}
}
\value{
A \code{lattice} plot
}
\description{
Accept a \code{survfit} object, and plot 'dose-survival curve' in the form
of a complementary cumulative distribution function of dose tolerability.
}
\seealso{
\code{\link{ds.curve}}
}
\author{
David C. Norris
}
\keyword{hplot}
\keyword{methods}
