% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{DE-class}
\alias{DE-class}
\title{An S4 class for simulating dose-titration study designs}
\description{
An S4 class for simulating dose-titration study designs
}
\section{Slots}{

\describe{
\item{\code{doses}}{A numeric vector of prospectively-determined discrete doses to trial.}

\item{\code{units}}{A string indicating dose units, e.g. \code{"mg/kg"}.}

\item{\code{MTDi}}{A numeric vector of optimal doses for simulated study participants.
Optionally a call to an \code{r<distribution>(...)} function which may be parsed
to calculated the \code{mtd_quantiles} slot.}

\item{\code{mtd_quantiles}}{A numeric vector of quantiles of the distribution
from which the MTDi slot was simulated. Intended mainly to support
visualization of this distribution, e.g. as an transparent overlay
on the dose-survival plot. NULL in case \code{MTDi} is provided verbatim.}

\item{\code{data}}{A data.frame with columns \strong{TODO}.}

\item{\code{stop_esc}}{integer \strong{TODO}.}

\item{\code{ds_conf_level}}{numeric \strong{TODO}.}

\item{\code{dose_drop_threshold}}{numeric \strong{TODO}.}

\item{\code{stop_esc_under}}{numeric \strong{TODO}.}

\item{\code{undo_esc_under}}{numeric \strong{TODO}.}
}}

