% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAT-package.R
\docType{data}
\name{de.bioRxiv.240846}
\alias{de.bioRxiv.240846}
\title{Simulated '3+3/PC' dose-titration study from bioRxiv paper no. 240846}
\format{The format is: List of 10 $ :'data.frame': 3 obs. of 4 variables:
..$ id : int [1:3] 1 2 3 ..$ period: num [1:3] 1 1 1 ..$ dose : num [1:3] 1
1 1 ..$ dlt : logi [1:3] FALSE FALSE FALSE $ :'data.frame': 9 obs. of 4
variables: ..$ id : int [1:9] 1 2 3 4 5 6 1 2 3 ..$ period: num [1:9] 1 1 1
2 2 2 2 2 2 ..$ dose : num [1:9] 1 1 1 1 1 1 2 2 2 ..$ dlt : logi [1:9]
FALSE FALSE FALSE FALSE FALSE FALSE ...  $ :'data.frame': 18 obs. of 4
variables: ..$ id : int [1:18] 1 2 3 4 5 6 1 2 3 7 ...  ..$ period: num
[1:18] 1 1 1 2 2 2 2 2 2 3 ...  ..$ dose : num [1:18] 1 1 1 1 1 1 2 2 2 1
...  ..$ dlt : logi [1:18] FALSE FALSE FALSE FALSE FALSE FALSE ...  $
:'data.frame': 30 obs. of 4 variables: ..$ id : int [1:30] 1 2 3 4 5 6 1 2 3
7 ...  ..$ period: num [1:30] 1 1 1 2 2 2 2 2 2 3 ...  ..$ dose : num [1:30]
1 1 1 1 1 1 2 2 2 1 ...  ..$ dlt : logi [1:30] FALSE FALSE FALSE FALSE FALSE
FALSE ...  $ :'data.frame': 45 obs. of 4 variables: ..$ id : int [1:45] 1 2
3 4 5 6 1 2 3 7 ...  ..$ period: num [1:45] 1 1 1 2 2 2 2 2 2 3 ...  ..$
dose : num [1:45] 1 1 1 1 1 1 2 2 2 1 ...  ..$ dlt : logi [1:45] FALSE FALSE
FALSE FALSE FALSE FALSE ...  $ :'data.frame': 61 obs. of 4 variables: ..$ id
: int [1:61] 1 2 3 4 5 6 1 2 3 7 ...  ..$ period: num [1:61] 1 1 1 2 2 2 2 2
2 3 ...  ..$ dose : num [1:61] 1 1 1 1 1 1 2 2 2 1 ...  ..$ dlt : logi
[1:61] FALSE FALSE FALSE FALSE FALSE FALSE ...  $ :'data.frame': 77 obs. of
4 variables: ..$ id : int [1:77] 1 2 3 4 5 6 1 2 3 7 ...  ..$ period: num
[1:77] 1 1 1 2 2 2 2 2 2 3 ...  ..$ dose : num [1:77] 1 1 1 1 1 1 2 2 2 1
...  ..$ dlt : logi [1:77] FALSE FALSE FALSE FALSE FALSE FALSE ...  $
:'data.frame': 89 obs. of 4 variables: ..$ id : int [1:89] 1 2 3 4 5 6 1 2 3
7 ...  ..$ period: num [1:89] 1 1 1 2 2 2 2 2 2 3 ...  ..$ dose : num [1:89]
1 1 1 1 1 1 2 2 2 1 ...  ..$ dlt : logi [1:89] FALSE FALSE FALSE FALSE FALSE
FALSE ...  ..- attr(\emph{, "stop.esc")= num 7 $ :'data.frame': 99 obs. of 4
variables: ..$ id : int [1:99] 1 2 3 4 5 6 1 2 3 7 ...  ..$ period: num
[1:99] 1 1 1 2 2 2 2 2 2 3 ...  ..$ dose : num [1:99] 1 1 1 1 1 1 2 2 2 1
...  ..$ dlt : logi [1:99] FALSE FALSE FALSE FALSE FALSE FALSE ...  ..-
attr(}, "stop.esc")= num 7 $ :'data.frame': 104 obs. of 4 variables: ..$ id
: int [1:104] 1 2 3 4 5 6 1 2 3 7 ...  ..$ period: num [1:104] 1 1 1 2 2 2 2
2 2 3 ...  ..$ dose : num [1:104] 1 1 1 1 1 1 2 2 2 1 ...  ..$ dlt : logi
[1:104] FALSE FALSE FALSE FALSE FALSE FALSE ...  ..- attr(*, "stop.esc")=
num 7}
\description{
This is a length-10 list
}
\examples{

data(de.bioRxiv.240846)
## maybe OXDSplot(last(de))

}
\references{
Norris DC. Precautionary Coherence Unravels Dose Escalation
Designs. bioRxiv. December 2017:240846. doi:10.1101/240846.
\url{https://www.biorxiv.org/content/early/2017/12/29/240846}
}
\keyword{datasets}
