% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose.survival.R
\name{dose.survival}
\alias{dose.survival}
\title{Extract interval-censored dose tolerance data from a dose titration study}
\usage{
dose.survival(de)
}
\arguments{
\item{de}{A data frame describing a dose-titration study}
}
\value{
A \code{Surv} object of type='interval2'
}
\description{
Extract interval-censored dose tolerance data from a dose titration study
}
\seealso{
\code{\link{dose.survfit}}
}
\author{
David C. Norris
}
\keyword{survival}
