% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAT-package.R
\docType{package}
\name{DTAT-package}
\alias{DTAT-package}
\alias{DTAT}
\title{c("\Sexpr[results=rd,stage=build]{tools:::Rd_package_title("DTAT")}",
"DTAT")\Sexpr{tools:::Rd_package_title("DTAT")}}
\description{
c("\Sexpr[results=rd,stage=build]{tools:::Rd_package_description("DTAT")}",
"DTAT")\Sexpr{tools:::Rd_package_description("DTAT")}
}
\details{
The DESCRIPTION file:
c("\Sexpr[results=rd,stage=build]{tools:::Rd_package_DESCRIPTION("DTAT")}",
"DTAT")\Sexpr{tools:::Rd_package_DESCRIPTION("DTAT")}
c("\Sexpr[results=rd,stage=build]{tools:::Rd_package_indices("DTAT")}",
"DTAT")\Sexpr{tools:::Rd_package_indices("DTAT")} ~~ An overview of how to
use the package, including the most important functions ~~
}
\references{
Norris DC. Dose Titration Algorithm Tuning (DTAT) should
supersede 'the' Maximum Tolerated Dose (MTD) in oncology dose-finding
trials. F1000Research. 2017;6:112. doi:10.12688/f1000research.10624.3.
\url{https://f1000research.com/articles/6-112/v3}

Norris DC. Costing 'the' MTD. bioRxiv. August 2017:150821.
doi:10.1101/150821.
\url{http://www.biorxiv.org/content/early/2017/08/22/150821}

Norris DC. Precautionary Coherence Unravels Dose Escalation Designs.
bioRxiv. December 2017:240846. doi:10.1101/240846.
\url{https://www.biorxiv.org/content/early/2017/12/29/240846}
}
\seealso{
\code{\link[pomp:pomp-package]{pomp}}
}
\author{
c("\Sexpr[results=rd,stage=build]{tools:::Rd_package_author("DTAT")}",
"DTAT")\Sexpr{tools:::Rd_package_author("DTAT")}

Maintainer:
c("\Sexpr[results=rd,stage=build]{tools:::Rd_package_maintainer("DTAT")}",
"DTAT")\Sexpr{tools:::Rd_package_maintainer("DTAT")}
}
