\name{DTAT-package}
\alias{DTAT-package}
\alias{DTAT}
\docType{package}
\title{
\packageTitle{DTAT}
}
\description{
\packageDescription{DTAT}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DTAT}
\packageIndices{DTAT}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{DTAT}

Maintainer: \packageMaintainer{DTAT}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
