\name{DTAT-package}
\alias{DTAT-package}
\alias{DTAT}
\docType{package}
\title{
\packageTitle{DTAT}
}
\description{
\packageDescription{DTAT}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DTAT}
\packageIndices{DTAT}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{DTAT}

Maintainer: \packageMaintainer{DTAT}
}
\references{
Norris DC. Dose Titration Algorithm Tuning (DTAT) should supersede 'the' Maximum Tolerated Dose (MTD) in oncology dose-finding trials. F1000Research. 2017;6:112. doi:10.12688/f1000research.10624.3. \url{https://f1000research.com/articles/6-112/v3}

Norris DC. Costing 'the' MTD. bioRxiv. August 2017:150821. doi:10.1101/150821. \url{http://www.biorxiv.org/content/early/2017/08/22/150821}

Norris DC. Precautionary Coherence Unravels Dose Escalation Designs. bioRxiv. December 2017:240846. doi:10.1101/240846. \url{https://www.biorxiv.org/content/early/2017/12/29/240846}

}
\seealso{
\code{\link[pomp:pomp-package]{pomp}}
}
