\name{DTAT-package}
\alias{DTAT-package}
\alias{DTAT}
\docType{package}
\title{
\packageTitle{DTAT}
}
\description{
\packageDescription{DTAT}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DTAT}
\packageIndices{DTAT}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{DTAT}

Maintainer: \packageMaintainer{DTAT}
}
\references{
Norris DC. Dose Titration Algorithm Tuning (DTAT) should supersede 'the' Maximum Tolerated Dose (MTD) in oncology dose-finding trials. F1000Research. 2017;6:112. doi:10.12688/f1000research.10624.2.
}
\seealso{
\code{\link[pomp:pomp-package]{pomp}}
}
\examples{
}
