% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.de.R
\name{plot.de}
\alias{plot.de}
\title{D3 visualization for dose-escalation/titration studies.}
\usage{
\method{plot}{de}(x, viewer = c("internal", "external", "browser"), ...)
}
\arguments{
\item{x}{The 'de' object to be transformed}

\item{viewer}{"internal" to use the RStudio internal viewer pane for output;
"external" to display in an external RStudio window; "browser" to display in
an external browser.}

\item{\dots}{Additional parameters passed to \code{r2d3}}
}
\value{
No value is returned. A plot is displayed.
}
\description{
An interactive visualization, including many mouseover interactions and
clickable dividers between the DLT assessment periods.
}
\details{
TODO: Detail all interactions supported, and their significance.
}
\examples{

\donttest{
library(DTAT)
de <- de.sim(testing=TRUE)
plot(de)
}

}
\author{
David C. Norris
}
\keyword{dynamic}
\keyword{hplot}
