% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaled.R
\name{scaled}
\alias{scaled}
\title{Power-law scaling for doses}
\usage{
scaled(dose, a = 4)
}
\arguments{
\item{dose}{A numeric vector of doses}

\item{a}{A numeric exponent for power-law rescaling}
}
\value{
A rescaled vector of doses
}
\description{
Implement an inverse power-law scaling for drug dose.
}
\author{
David C. Norris
}
