% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titrate.R
\name{titrate}
\alias{titrate}
\title{Perform Dose Titration Algorithm Tuning (DTAT)}
\usage{
titrate(draw.days = NULL, Tcyc = 3 * 7 * 24, Ncycles = 10,
  doserange = c(50, 500), dta = NULL)
}
\arguments{
\item{draw.days}{TODO: Describe \code{draw.days} here~~}

\item{Tcyc}{TODO: Describe \code{Tcyc} here~~}

\item{Ncycles}{TODO: Describe \code{Ncycles} here~~}

\item{doserange}{TODO: Describe \code{doserange} here~~}

\item{dta}{A Dose Titration Algorithm (DTA) to drive the titration}
}
\value{
A list with 2 components:
\item{course}{A data frame containing cycle-wise measures
of each id's titration course}
\item{anc.ts}{A data frame detailing hourly ANC measures for each id}
}
\description{
Perform neutrophil-guided dose titration of a chemotherapy drug.
}
\note{
TODO: Once I've implemented several nontrivial dosing algorithms, I
should consider whether the 'doserange' case ought to be handled by a
special non-adaptive DTA.
}
\author{
David C. Norris
}
